requirements.txt
tensorflow==2.16.1
keras==3.3.3
numpy==1.26.4
pillow==10.4.0
opencv-python==4.10.0.84
matplotlib==3.9.2
scikit-learn==1.5.2
